Move randomNormallyDistributedFloat into distribution.js.
